Move TxScratchpad tests to tx_scratchpad.rs
Moved test_cache_warmup_changeset, test_cache_warmup_overrides, and all
helper functions to the tx_scratchpad.rs file. Removed the now-empty test
module from mod.rs, completing the test reorganization.
